improve - use long replace int

+++ craft-atom-redis/src/main/java/org/craft/atom/redis/DefaultMasterSlaveRedis.java
@@ -237,7 +237,7 @@ public class DefaultMasterSlaveRedis implements MasterSlaveRedis {
 	}
 
 	@Override
-	public Long pexpire(String key, int milliseconds) {
+	public Long pexpire(String key, long milliseconds) {
 		return master().pexpire(key, milliseconds);
 	}
